fix(BE): location lat long parameters converted to float

+++ src/main/kotlin/com/ramazanm/saveloc/data/dto/PlaceDTO.kt
@@ -5,8 +5,8 @@ import com.ramazanm.saveloc.data.model.Place
 data class PlaceDTO(
     val name: String,
     val description: String,
-    val lat: Long,
-    val long: Long,
+    val lat: Float,
+    val long: Float,
     val photos: List<String>
 ) {
     fun toPlaceEntity(userId: String) = Place(
